requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
blinker==1.8.2
brotli==1.2.0
cachetools==6.2.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.2
coverage==7.11.3
cryptography==44.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
durationpy==0.10
fastapi==0.128.0
Flask==3.0.3
Flask-Login==0.6.3
flatbuffers==25.2.10
flex==6.14.1
fonttools==4.57.0
fsspec==2026.1.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.63.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
ibm-cloud-sdk-core==3.23.0
ibm-platform-services==0.64.0
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpointer==3.0.0
keras==3.12.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
namex==0.1.0
nodemon==0.0.16
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
orjson==3.10.6
packaging==24.1
pandas==2.3.3
pbr==6.1.1
pillow==11.2.1
pluggy==1.6.0
postgrest==2.27.2
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.1.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.1.2
pypdf==6.6.2
PyPDF2==3.0.1
pypdfium2==5.3.0
PyPika==0.50.0
pyroaring==1.0.3
pyspnego==0.11.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pywin32==311
PyYAML==6.0.1
qiskit==2.0.0
qiskit-ibm-runtime==0.38.0
QsKit==0.0.1
realtime==2.27.2
requests==2.32.3
requests_ntlm==1.3.0
rfc3987==1.3.8
rich==14.2.0
rsa==4.9.1
rustworkx==0.16.0
scipy==1.15.2
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sspilib==0.3.1
starlette==0.50.0
stevedore==5.4.1
storage3==2.27.2
StrEnum==0.4.15
strict-rfc3339==0.7
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
symengine==0.13.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.40.0
validate_email==1.3
watchdog==6.0.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.0.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
